requirements.txt
flask
Authlib
requests
python-dotenv
Flask-Session
msal
pyjwt
identity
werkzeug
pytz
APScheduler
python-dateutil
Flask-SQLAlchemy